Split up compiler pipeline phases
Move interface Monitors to frontend.
This makes the first two compiler pipelines use BaseContext only.
