+++ community/cypher/cypher-compiler-3.2/src/main/scala/org/neo4j/cypher/internal/compiler/v3_2/ast/rewriters/rewriteEqualityToInPredicate.scala
@@ -19,17 +19,18 @@
  */
 package org.neo4j.cypher.internal.compiler.v3_2.ast.rewriters
 
+import org.neo4j.cypher.internal.compiler.v3_2.phases.{CompilationState, Condition, Phase}
+import org.neo4j.cypher.internal.frontend.v3_2.ast._
+import org.neo4j.cypher.internal.frontend.v3_2.phases.BaseContext
 import org.neo4j.cypher.internal.frontend.v3_2.phases.CompilationPhaseTracer.CompilationPhase
 import org.neo4j.cypher.internal.frontend.v3_2.phases.CompilationPhaseTracer.CompilationPhase.AST_REWRITE
-import org.neo4j.cypher.internal.compiler.v3_2.phases.{CompilationState, Condition, CompilerContext, Phase}
-import org.neo4j.cypher.internal.frontend.v3_2.ast._
 import org.neo4j.cypher.internal.frontend.v3_2.{Rewriter, bottomUp}
 
 case object rewriteEqualityToInPredicate extends StatementRewriter {
 
   override def description: String = "normalize equality predicates into IN comparisons"
 
-  override def instance(ignored: CompilerContext): Rewriter = bottomUp(Rewriter.lift {
+  override def instance(ignored: BaseContext): Rewriter = bottomUp(Rewriter.lift {
     // id(a) = value => id(a) IN [value]
     case predicate@Equals(func@FunctionInvocation(_, _, _, IndexedSeq(idExpr)), idValueExpr)
       if func.function == functions.Id =>
@@ -47,12 +48,12 @@ case object rewriteEqualityToInPredicate extends StatementRewriter {
   override def postConditions: Set[Condition] = Set.empty
 }
 
-trait StatementRewriter extends Phase[CompilerContext] {
+trait StatementRewriter extends Phase[BaseContext] {
   override def phase: CompilationPhase = AST_REWRITE
 
-  def instance(context: CompilerContext): Rewriter
+  def instance(context: BaseContext): Rewriter
 
-  override def process(from: CompilationState, context: CompilerContext): CompilationState = {
+  override def process(from: CompilationState, context: BaseContext): CompilationState = {
     val rewritten = from.statement.endoRewrite(instance(context))
     from.copy(maybeStatement = Some(rewritten))
   }
