+++ community/cypher/cypher-compiler-3.2/src/main/scala/org/neo4j/cypher/internal/compiler/v3_2/phases/AstRewriting.scala
@@ -63,7 +63,7 @@ case class AstRewriting(sequencer: String => RewriterStepSequencer, shouldExtrac
 }
 
 object LateAstRewriting extends StatementRewriter {
-  override def instance(context: CompilerContext): Rewriter = inSequence(
+  override def instance(context: BaseContext): Rewriter = inSequence(
     collapseMultipleInPredicates,
     nameUpdatingClauses,
     projectNamedPaths,
